# GetEnvironmentVariable: Avoid StringBuilder marshaling overhead