Try a stack allocated buffer first, then fallback to using the shared array pool.

Results on my machine:


|             Method |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------- |----------:|----------:|----------:|-------:|----------:|
|  `"TMP"` Before | 156.51 ns | 2.0132 ns | 1.8831 ns | 0.0875 |     368 B |
|  `"TMP"` After |  92.88 ns | 0.5002 ns | 0.4434 ns | 0.0209 |      88 B |
| `"PATH"` Before | 621.59 ns | 2.2668 ns | 2.1204 ns | 0.9565 |    4016 B |
| `"PATH"` After | 287.02 ns | 1.0377 ns | 0.8666 ns | 0.2704 |    1136 B |


`"TMP"` is under the stack limit and `"PATH"` is over the limit (uses the pool).