@justinvp 2017-10-14
@dotnet-bot test this please