@danmosemsft 2017-10-15
@dotnet-bot test Windows_NT x86 full_opt legacy_backend CoreCLR Perf Tests Correctness please (nolog)