@justinvp 2017-10-16
I had added `clearArray: true` when returning the array to the pool out of caution and it didn't appear to significantly impact the perf, but re-running the microbenchmarks again does show a slowdown.

`clearArray: true`:

|                Method |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
|---------------------- |----------:|----------:|----------:|-------:|----------:|
|    Before | 702.23 ns | 2.0626 ns | 1.8284 ns | 1.1282 |    4736 B |
|    After | 723.63 ns | 2.8812 ns | 2.5541 ns | 0.3271 |    1376 B |

`clearArray: false`:

|                Method |      Mean |      Error |     StdDev |  Gen 0 | Allocated |
|---------------------- |----------:|-----------:|-----------:|-------:|----------:|
|    Before | 684.89 ns |  3.7051 ns |  3.2845 ns | 1.1282 |    4736 B |
|    After | 316.44 ns |  1.9137 ns |  1.7901 ns | 0.3276 |    1376 B |

Thus, I've added a commit that removes `clearArray: true` since:

 - Any code in the process can already read environment variables whenever it wants
 - Clearing the array before returning it to the pool is not done anywhere else in corelib
 - CoreRT's `ExpandEnvironmentVariables` implementation uses the pool and [doesn't clear the array](https://github.com/dotnet/corert/blob/6a9700085acec28d32d14ab89bcf9cd91ff1e581/src/System.Private.CoreLib/src/System/Environment.Windows.cs#L46)

Let me know if you have any concerns